Add TLSVerificationModal component for managing TLS verification settings. This modal allows users to toggle TLS verification with appropriate warnings about potential security risks and service interruptions. Refactor TLSVerificationUpdater to utilize the new modal, improving code organization and user experience.
